Add Functions and Classes to respective modules
