+++ modules/module-4.js
@@ -1,0 +1,6 @@
+const toggleVisbility = (activeEle) => {
+  activeEle.classList.toggle('active');
+  activeEle.classList.remove('hide-class');
+}
+
+export {toggleVisbility};
